model/requirements.txt
numpy
pandas
opencv-python
torch
scipy
Pillow